Comments, Raw Data Edit, and Missing Question Record Data
This commit adds comments to the clean_data script for clarity. It corrects a free response in the raw data that was showing up as #NAME because it was entered with an "=" causing Excel to treat it as a formula. It also exports a list of records that require data to be populated for questions that were added to the survey shortly after data collection was started.