Functions! And other code tweaks
This commit contains some major changes to the .RProfile file. First, a file for two functions is added. One of the functions imports GTFS data when you pass the name of folder in which the data is stored. This reduces the amount of code on the .RProfile script because of another change. That change is the incorporation of the full Dec 2020 GTFS export into the .RProfile script. Both the Dec 2019 GTFS data and the Dec 2020 GTFS data are used to find the street names, routes, and directions for all stop IDs from the Bus Stop Census.

In the 01_clean_data.R script,  the major changes are adding code to import corrections from Google Sheets, using a function to apply corrections to the data, and adding code to deal with duplicates.