More Contradictory Code Updates and File Exports
This commit contains an export for the instance where the surveyor selects no seating while also selecting that the stop had a dirty seating area.

The commit also added the code for the corrections for the contradictory answers as well as minor cosmetic changes to the structure of the code blocks.